+++ src/lib/utils/entry-point.ts
@@ -9,11 +9,10 @@ import {
     ignorePackage,
     loadPackageManifest,
 } from "./package-manifest";
-import { createMinimatch, matchesAny, nicePath } from "./paths";
+import { createMinimatch, matchesAny, nicePath, normalizePath } from "./paths";
 import type { Logger } from "./loggers";
 import type { Options } from "./options";
-import { getCommonDirectory, glob, normalizePath } from "./fs";
-import { filterMap } from "./array";
+import { deriveRootDir, glob } from "./fs";
 import { assertNever } from "./general";
 
 /**
@@ -180,8 +179,7 @@ function getEntryPointsForPaths(
     options: Options,
     programs = getEntryPrograms(logger, options)
 ): DocumentationEntryPoint[] | undefined {
-    const baseDir =
-        options.getValue("basePath") || getCommonDirectory(inputFiles);
+    const baseDir = options.getValue("basePath") || deriveRootDir(inputFiles);
     const entryPoints: DocumentationEntryPoint[] = [];
 
     entryLoop: for (const fileOrDir of inputFiles.map(normalizePath)) {
@@ -237,7 +235,7 @@ export function getExpandedEntryPointsForPaths(
 }
 
 function expandGlobs(inputFiles: string[]) {
-    const base = getCommonDirectory(inputFiles);
+    const base = deriveRootDir(inputFiles);
     const result = inputFiles.flatMap((entry) =>
         glob(entry, base, { includeDirectories: true, followSymlinks: true })
     );
@@ -339,18 +337,6 @@ function expandInputFiles(
     });
 
     return files;
-}
-
-function deriveRootDir(packageGlobPaths: string[]): string {
-    const globs = createMinimatch(packageGlobPaths);
-    const rootPaths = globs.flatMap((glob) =>
-        filterMap(glob.set, (set) => {
-            const stop = set.findIndex((part) => typeof part !== "string");
-            const path = stop === -1 ? set : set.slice(0, stop);
-            return `/${path.join("/")}`;
-        })
-    );
-    return getCommonDirectory(rootPaths);
 }
 
 /**